feat: image compressor tool
