+++ src/screens/image-ocr/compressor/image-ocr-compressor.tsx
@@ -1,0 +1,3 @@
+import { Typography } from "antd";
+
+export const ImageOcrCompressor = () => <Typography.Paragraph>Image compressor coming soon.</Typography.Paragraph>;
