+++ src/screens/image-ocr/image-ocr.constants.ts
@@ -1,0 +1,4 @@
+export const TAB_ITEMS = [
+  { key: "/image-ocr/ocr", label: "OCR" },
+  { key: "/image-ocr/compressor", label: "Compressor" },
+];
